minor wording change
For exercises where we are having learners assume they have a particular working directory, I prefer to say "if your directory is . . . " rather than "this is your working directory" because I think learners might be confused by that wording and think that is actually their working directory (by definition). This is a minor thing, but I think it improves the clarity a little. Otherwise no changes to this episode!